+++ designer/server/src/main/scala/pl/touk/nussknacker/ui/process/exception/ProcessIllegalAction.scala
@@ -3,7 +3,7 @@ package pl.touk.nussknacker.ui.process.exception
 import pl.touk.nussknacker.engine.api.deployment.{ScenarioActionName, StateStatus}
 import pl.touk.nussknacker.engine.api.process.ProcessName
 import pl.touk.nussknacker.ui.IllegalOperationError
-import pl.touk.nussknacker.ui.process.scenariostatus.ScenarioStatusWithAllowedActions
+import pl.touk.nussknacker.ui.process.deployment.scenariostatus.ScenarioStatusWithAllowedActions
 
 final case class ProcessIllegalAction(message: String) extends IllegalOperationError(message, details = "")
