+++ designer/server/src/test/scala/pl/touk/nussknacker/ui/process/scenariostatus/InconsistentStateDetectorTest.scala
@@ -6,6 +6,7 @@ import pl.touk.nussknacker.engine.api.deployment.StatusDetails
 import pl.touk.nussknacker.engine.api.deployment.simple.SimpleStateStatus
 import pl.touk.nussknacker.engine.api.deployment.simple.SimpleStateStatus.ProblemStateStatus
 import pl.touk.nussknacker.engine.deployment.DeploymentId
+import pl.touk.nussknacker.ui.process.deployment.scenariostatus.InconsistentStateDetector
 
 import java.util.UUID
